requirements.txt
Flask==2.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
keras==2.6.0
numpy==1.19.5
Pillow==8.4.0
tensorflow-cpu==2.6.2
